python-flask-backEnd/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2
virtualenv
python-dotenv
python-decouple
google-cloud-bigquery
pandas
flask-cors
flask-socketio